requirements.txt
# for production and local builds
# > python -m venv venv --prompt="name-prompt-here"
# > venv/Scripts/activate
# > pip install -r requirements.txt
# > deactivate

certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.10.0
fonttools==4.41.0
idna==2.10
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.1
schedule==1.2.0
six==1.16.0
supervision==0.11.1
tqdm==4.65.0
urllib3==2.0.4
wget==3.2
